pyproject.toml
[tool.poetry]
name = "python-gql-compiler"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gql = "3.0.0rc0"
PyYAML = "^6.0"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
flake8 = "^4.0.1"
isort = "^5.10.1"
black = "^22.1.0"
mypy-extensions = "^0.4.3"
pyproject-flake8 = "^0.0.1-alpha.2"
gql = {version = "3.0.0rc0", extras = ["all"]}
strawberry-graphql = {extras = ["debug-server"], version = "^0.96.0"}
types-PyYAML = "^6.0.4"
types-requests = "^2.27.8"
pre-commit = "^2.17.0"
coverage = "^6.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110

[tool.isort]
line_length = 110
multi_line_output = 3
include_trailing_comma = true

[tool.flake8]
max-line-length = 110
max-complexity = 18
extend-ignore = "E203,"  # blackと競合してる

[tool.mypy]
plugins = ["strawberry.ext.mypy_plugin"]
namespace_packages = true
ignore_missing_imports = true  # スタブがないライブラリを無視する
